sbom-tracer/requirements.txt
setuptools
hyperframe
hpack
six
PyYAML
click